Create LICENSE.md
I added the copyright statement and the dual licenses for the book's
content. The main content for the book (text and graphics) are licensed
under the Creative Commons Attribution-NoDerivatives 4.0 International
License. The source code that is produced for the book is licensed under
the MIT License. The MIT License will allow the source code to be used
for both non-commercial and commercial purposes with attribution.