Create README.md
I started the content for the README.md document. I added a description
of the repository, the book, and explained why I am writing the book. I
added a section to explain the licensing. I added a section about
accepting contributions to the book.

I created a placeholder CONTRIBUTING.md document. I will add more
information to this document in the future to explain how to contribute
new content or samples to the book.