Create the book
I used the `mdbook init` command to generate the starter files for the
book.

I created the serve.sh shell program in the root directory of the
project. When testing out the `mdbook serve` command in the development
container, I wasn't able to view the book content. I discovered that the
mdbook web server by default listens on an IPv6 address and I was not
able to connect from the host. I created serve.sh to run the web server
listening on 127.0.0.1 to force it to IPv4.