Generate ePUB version of the book
I added mdbook-epub to the project. mdbook-epub will extend mdbook to
generate an ePUB version of the book that can be installed in a reader
program such as Apple Books on the iPad or macOS.

I created a custom development container feature to install mdbook-epub
and added it to the development container.